# backend: de-virtualize regalloc 