… internal/engine/wazevo/backend/backend_test.go
… internal/engine/wazevo/backend/executable_context.go
… internal/engine/wazevo/backend/isa/amd64/abi_entry_preamble_test.go
… internal/engine/wazevo/backend/isa/amd64/abi_go_call_test.go
… internal/engine/wazevo/backend/isa/amd64/instr_encoding_test.go
… internal/engine/wazevo/backend/isa/amd64/machine_pro_epi_logue_test.go
… internal/engine/wazevo/backend/isa/amd64/machine_regalloc_test.go
… internal/engine/wazevo/backend/isa/amd64/machine_test.go
… internal/engine/wazevo/backend/isa/amd64/util_test.go
… internal/engine/wazevo/backend/isa/arm64/abi_entry_preamble_test.go
… internal/engine/wazevo/backend/isa/arm64/abi_go_call.go
… internal/engine/wazevo/backend/isa/arm64/abi_go_call_test.go
… internal/engine/wazevo/backend/isa/arm64/lower_instr.go
… internal/engine/wazevo/backend/isa/arm64/lower_instr_test.go
… internal/engine/wazevo/backend/isa/arm64/machine.go
… internal/engine/wazevo/backend/isa/arm64/machine_pro_epi_logue.go
… internal/engine/wazevo/backend/isa/arm64/machine_pro_epi_logue_test.go
… internal/engine/wazevo/backend/isa/arm64/machine_regalloc.go
… internal/engine/wazevo/backend/isa/arm64/machine_regalloc_test.go
… internal/engine/wazevo/backend/isa/arm64/machine_test.go
… internal/engine/wazevo/backend/isa/arm64/util_test.go
… internal/engine/wazevo/backend/machine.go
… internal/engine/wazevo/backend/machine_test.go
… internal/engine/wazevo/backend/regalloc.go
… internal/engine/wazevo/backend/regalloc/api.go
… internal/engine/wazevo/backend/regalloc/api_test.go
… internal/engine/wazevo/backend/regalloc/regalloc.go
… internal/engine/wazevo/backend/regalloc/regalloc_test.go
… internal/engine/wazevo/backend/regalloc/regset.go