This de-virtualize the regalloc code, and we got the overall 5% 
compilation perf improvements without runtime impact.

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero
                      │  old.txt   │             new.txt              │
                      │   sec/op   │   sec/op    vs base              │
Compilation/wazero-10   1.722 ± 2%   1.636 ± 1%  -4.95% (p=0.002 n=6)
Compilation/zig-10      3.761 ± 0%   3.589 ± 0%  -4.58% (p=0.002 n=6)
Compilation/zz-10       16.03 ± 0%   15.20 ± 1%  -5.15% (p=0.002 n=6)
geomean                 4.700        4.470       -4.90%

                      │   old.txt    │              new.txt               │
                      │     B/op     │     B/op      vs base              │
Compilation/wazero-10   286.8Mi ± 0%   271.2Mi ± 0%  -5.44% (p=0.002 n=6)
Compilation/zig-10      601.7Mi ± 0%   596.2Mi ± 0%  -0.90% (p=0.002 n=6)
Compilation/zz-10       538.1Mi ± 0%   528.9Mi ± 0%  -1.72% (p=0.002 n=6)
geomean                 452.8Mi        440.6Mi       -2.71%

                      │   old.txt   │              new.txt              │
                      │  allocs/op  │  allocs/op   vs base              │
Compilation/wazero-10   448.4k ± 0%   448.4k ± 0%       ~ (p=0.699 n=6)
Compilation/zig-10      274.6k ± 0%   274.6k ± 0%       ~ (p=0.974 n=6)
Compilation/zz-10       618.3k ± 0%   618.5k ± 0%       ~ (p=0.065 n=6)
geomean                 423.8k        423.9k       +0.02%
```

#2182