@mathetake 2024-06-24
need to fix amd64 yet